+++ CHANGELOG.rst
@@ -12,10 +12,11 @@ Unreleased
 
 Added
 -----
-- Vector3d.scatter() plot method plotting the vectors as unit vectors in the
-  stereographic projection
+- Vector3d.scatter() plot method showing vectors as unit vectors in the stereographic
+  projection
 - User guide with Jupyter notebooks as part of the Read the Docs documentation
-- Stereographic projection using Matplotlib's projections framework
+- Stereographic projection using Matplotlib's projections framework for plotting
+  vectors, great and small circles, and symmetry elements
 - orix.projections module for projecting vectors to various coordinates, including
   stereographic coordinates
 - .ang file writer for CrystalMap objects (via orix.io.save())